# internal/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go + PocketBase backend for QuantumAtlas: paper\ncollection + search + database. Write endpoints and\nthe search/paper read surface require a bearer token\n(PAT or PocketBase session) plus the matching scope.\nSee the auth model docs for the scope vocabulary.",
        "title": "QuantumAtlas API",
        "contact": {
            "name": "QuantumAtlas",
            "url": "https://quantum-atlas.ai"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/downloader/remote-jobs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "PostgreSQL-persisted snapshot of outbound fleet tasks (most\nrecently updated first, capped at 500): id, worker_id, state\n(queued/running/staged/done/failed), identifier and error.\nUnlike /api/downloader/jobs this survives server restarts.\nAnswers 200 with {enabled: false, jobs: []} when the remote\nfleet is disabled. Requires papers:read.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Downloader"
                ],
                "summary": "Downloader remote jobs",
                "responses": {
                    "200": {
                        "description": "{enabled: bool, jobs: [{id, worker_id, state, identifier, error?, updated_at}]}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the signed-in user's profile:\n{id, email, name, avatar, github_login, gitea_login,\ngithub_bound, gitea_bound, is_admin, is_superadmin,\ncreated}.\navatar is the PocketBase file name — build the URL via the\nPocketBase files API. *_bound reports which OAuth2\nidentities are linked to the record (the dashboard's\naccount-binding panels).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Me"
                ],
                "summary": "My profile",
                "responses": {
                    "200": {
                        "description": "{id, email, name, avatar, github_login, gitea_login, github_bound, gitea_bound, is_admin, is_superadmin, created}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "PAT auth not accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/me/usage": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns {metric, today, limit, llm_tokens} for the\nagentic-search surface: today's call count, the caller's\neffective daily limit (per-user override \u003e plan \u003e\nsearch.agentic.daily_limit) and the LLM tokens consumed\ntoday. Same metering shape as the agentic response's usage\nblock.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Me"
                ],
                "summary": "My usage",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.meUsageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "PAT auth not accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "postgres registry unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "routes.meUsageResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "llm_tokens": {
                    "type": "integer"
                },
                "metric": {
                    "type": "string"
                },
                "today": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer \u003ctoken\u003e\" — two credential shapes\naccepted: a Personal Access Token\n(`Authorization: Bearer qat_...`, minted at\n`/pat` after GitHub OAuth login), or the\nconfig-loaded system PAT (set\n`system_pat.token` in config.yaml on the\nserver, send the plaintext as\n`Authorization: Bearer \u003cvalue\u003e`).\nBrowser callers are authenticated through\npb.authStore (no copy step) — only non-browser\ncallers need an explicit bearer.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}